Factor PythonPrinter out of inductor
[ghstack-poisoned]
